requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
Django==4.2.6
django-debug-toolbar==4.2.0
django-dotenv==1.4.2
django-redis==5.4.0
djangorestframework==3.14.0
pytz==2023.3.post1
redis==5.0.1
sqlparse==0.4.4
typing_extensions==4.8.0
